Little fix

+++ lottie-ios/Classes/Private/LOTAnimationView.m
@@ -367,7 +367,7 @@
     }
     _compLayer = [[LOTCompositionLayer alloc] initWithLayerGroup:_sceneModel.layerGroup
                                                   withAssetGroup:_sceneModel.assetGroup
-                                               withStartProgress:_sceneModel.startFrame.doubleValue / _sceneModel.framerate.doubleValue
+                                               withStartProgress:_sceneModel.startFrame.floatValue / _sceneModel.framerate.floatValue
                                                       withBounds:_sceneModel.compBounds
                                                         inBundle:_bundle];
     [_timingLayer addSublayer:_compLayer];
